requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
sqlalchemy[asyncio]==2.0.23
alembic==1.12.1
asyncpg==0.29.0
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP clients
httpx==0.25.2
aiohttp==3.9.1

# Data processing
pandas==2.1.3
numpy==1.25.2
python-dateutil==2.8.2

# Vector database
qdrant-client==1.7.0

# LLM and embeddings
openai==1.3.7
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.1.1

# Environment and config
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Data validation
validators==0.22.0
email-validator==2.1.0

# Scheduling
celery==5.3.4
APScheduler==3.10.4

# Document processing (for file extraction)
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
beautifulsoup4==4.12.2